Separated style files from the components
