+++ src/main/components/BulletList/BulletList.styl
@@ -1,0 +1,54 @@
+_bulletListImageDesktopNumberOfColumns =  NousRejoindre__$leftPartNumberOfColumns
+_bulletListImageDesktopPercentage = (_bulletListImageDesktopNumberOfColumns/layout__$gridSectionNumberOfColumns*100)
+_bulletListListDesktopPercentage = 100 - _bulletListImageDesktopPercentage
+
+_bulletListListItemPictoSize = BulletPoints__$smallBulletSize
+
+.BulletList
+  clearfix()
+
+.BulletList-image-wrapper
+  width 100%
+  .size-class-not-width-compact &
+    text-align right
+    float left
+    width (_bulletListImageDesktopPercentage) %
+
+.BulletList-image
+  width 100%
+  height 140px
+  background-size cover
+  display inline-block
+  .size-class-not-width-compact &
+    max-width 256px
+
+.BulletList-list
+  padding 30px 0 0 0
+  .size-class-not-width-compact &
+    padding 0
+    float left
+    width (_bulletListListDesktopPercentage)%
+
+.BulletList-list-item
+  clearfix()
+  padding-bottom 5px
+  & + &
+    padding-top 5px
+
+.BulletList-list-item-picto
+  width (_bulletListListItemPictoSize)px
+  float left
+  padding ((font__$textLineHeight - _bulletListListItemPictoSize)/2)px (BulletPoints__$smallBulletMobilePaddingHorizontal*2)px ((font__$textLineHeight - _bulletListListItemPictoSize)/2)px 0
+  .size-class-not-width-compact &
+    padding ((font__$textLineHeight - _bulletListListItemPictoSize)/2)px (BulletPoints__$smallBulletDesktopPaddingHorizontal)px
+
+.BulletList-list-item-text, .BulletList-bottom-text
+  width auto
+  color color__$text
+  padding-left (_bulletListListItemPictoSize+BulletPoints__$smallBulletMobilePaddingHorizontal*2)px
+  line-height (font__$textLineHeight)px
+  .size-class-not-width-compact &
+    padding-left (_bulletListListItemPictoSize+BulletPoints__$smallBulletDesktopPaddingHorizontal*2)px
+
+.BulletList-bottom-text
+  padding-top 20px
