+++ src/main/components/CTAButton/CTAButton.styl
@@ -1,0 +1,16 @@
+_CTAButtonPrimaryBackgroundColor = color__$flashOrange
+_CTAButtonSecondaryBackgroundColor = color__$lightBlue
+
+.CTAButton
+  display block
+  text-transform uppercase
+  color white
+  text-align center
+  font__useTextRegular font__$textFontSize
+  text-decoration none
+  padding 14px 42px
+  border-radius 5px
+  &.primary
+    background-color _CTAButtonPrimaryBackgroundColor
+  &.secondary
+    background-color _CTAButtonSecondaryBackgroundColor
