+++ src/main/components/CallToActionLayer/CallToActionLayer.styl
@@ -1,0 +1,19 @@
+.CallToActionLayer-button-outer-wrapper
+  padding 12px 0
+  min-width (284px / 2)
+
+.CallToActionLayer-button
+  .size-class-width-compact &
+    padding-left 10px !important
+    padding-right 10px !important
+
+.CallToActionLayer-logo
+  position relative
+  path
+    fill white
+  .size-class-width-compact &
+    padding 10px 0
+    height 45px
+    top 2px
+  .size-class-not-width-compact &
+    height 64px
