+++ src/main/components/CallToActionLayer/CallToActionLayer.vue
@@ -9,7 +9,7 @@
 <script>
 
 import AppHeaderSubBarLayer from 'xebia-web-common/components/AppHeaderSubBarLayer';
-import CTAButton from 'components/CTAButton';
+import CTAButton from 'components/CTAButton/CTAButton';
 
 import LogoXebiaVisMaVie from 'src/generated/components/LogoXebiaVisMaVie';
 
@@ -28,24 +28,4 @@ export default {
 };
 </script>
 
-<style lang="stylus">
-	.CallToActionLayer-button-outer-wrapper
-		padding 12px 0
-		min-width (284px / 2)
-
-	.CallToActionLayer-button
-		.size-class-width-compact &
-			padding-left 10px !important
-			padding-right 10px !important
-
-	.CallToActionLayer-logo
-		position relative
-		path
-			fill white
-		.size-class-width-compact &
-			padding 10px 0
-			height 45px
-			top 2px
-		.size-class-not-width-compact &
-			height 64px
-</style>
+<style lang="stylus" src="./CallToActionLayer.styl"></style>
