+++ src/main/components/StatCard/StatCard.styl
@@ -1,0 +1,46 @@
+_pictoColor = color__$orange
+_statColor = color__$blue
+_textColor = color__$blue
+
+_statFontSize = font__$cardTitleFontSize
+_textFontSize = font__$cardTextFontSize
+
+.StatCard
+  background-color white
+  border-radius 0 0 5px 5px
+
+.StatCard-inner-wrapper
+  padding 0 10px 10px 10px
+
+.StatCard-picto-wrapper
+  height (100*2/3) %
+  width 100%
+  padding-bottom 15px
+  border-bottom 1px solid color__$sectionEven
+
+.StatCard-picto
+  color _pictoColor
+  margin-left auto
+  margin-right auto
+  display block
+
+.StatCard-stat-text-wrapper
+  height (100*1/3) %
+
+.StatCard-stat
+  padding-top 10px
+  color _statColor
+  font__useTitleRegular _statFontSize
+  text-align center
+
+.StatCard-text
+  padding-top 5px
+  color _textColor
+  font__useTextLight _textFontSize
+  text-align center
+  height 32px
+  display table
+  width 100%
+  & span
+    display table-cell
+    vertical-align middle
