+++ src/main/components/StatCardList/StatCardList.styl
@@ -1,0 +1,8 @@
+.StatCard-list
+  clearfix()
+
+.size-class-width-compact
+  _grid('.StatCard', layout__$gridContainerNumberOfColumns, 2, 1, 3vw)
+
+.size-class-not-width-compact
+  _grid('.StatCard', layout__$gridContainerNumberOfColumns, 4, 2, 40px)
