+++ src/main/components/Tester.vue
@@ -3,23 +3,23 @@
 </template>
 
 <script>
-import OfferAccordion from 'components/OfferAccordion';
+  import OfferAccordion from 'components/OfferAccordion';
 
-import annonces from 'assets/data/nous-rejoindre/annonces.yaml';
+  import annonces from 'assets/data/nous-rejoindre/annonces.yaml';
 
-export default {
-	name: 'Tester',
-	data: function () {
-		return {
-		    jobs: annonces,
-		    talentButtonLabel: 'Vous avez d\'autres talents&nbsp;? C\'est&nbsp;par&nbsp;ici&nbsp;!',
-		    jobButtonLabel: 'Postuler !'
-		};
-	},
-	components: {
-	    OfferAccordion
-	}
-};
+  export default {
+    name : 'Tester',
+    data : function() {
+      return {
+        jobs : annonces,
+        talentButtonLabel : 'Vous avez d\'autres talents&nbsp;? C\'est&nbsp;par&nbsp;ici&nbsp;!',
+        jobButtonLabel : 'Postuler !'
+      };
+    },
+    components : {
+      OfferAccordion
+    }
+  };
 </script>
 
 <style lang="stylus">
